--- app/models/exam.rb
@@ -1,8 +1,0 @@
-class Exam < ApplicationRecord
-
-    belongs_to :pset
-    delegate :name, to: :pset
-
-    serialize :config, Hash
-
-end
