+++ plugins/exam/app/models/plugins/exam/exam.rb
@@ -1,0 +1,12 @@
+module Plugins
+  module Exam
+    class Exam < ApplicationRecord
+
+      belongs_to :pset
+      delegate :name, to: :pset
+
+      serialize :config, Hash
+
+    end
+  end
+end
