+++ plugins/exam/app/views/plugins/exam/admin/exams/index.html.erb
@@ -1,0 +1,13 @@
+<% content_for :title do %>
+<h1>Exam administration</h1>
+<% end %>
+
+<div class="list-group">
+<% @exams.all.each do |exam| %>
+<%= link_to exam.name.titleize, edit_admin_exam_path(exam), class: 'list-group-item list-group-item-action' %>
+<% end %>
+</div>
+
+<div class="list-group mt-5">
+<%= link_to "Code list", list_codes_admin_exams_path, class: 'list-group-item list-group-item-action', data: { turbo: false } %>
+</div>
