+++ plugins/exam/app/views/plugins/exam/exams/index.html.erb
@@ -1,0 +1,12 @@
+<section class="markdown" id="exams-content">
+    <h1>Tentamens</h1>
+<% @exams.each do |e| %>
+<p>
+    <%= button_to e.name.humanize, exam_url(id: e.id), method: :post,  class: 'btn btn-secondary', data: { turbo: false } %>
+    <% if e.locked? || e.pset.submits.any? &&
+        (e.pset.submits.first.grade.present? || e.pset.submits.first.locked?) %>
+    Locked
+    <% end %>
+</p>
+<% end %>
+</section>
