+++ plugins/exam/views/admin/exams/index.html.erb
